introduced resources access

+++ src/test/java/it/geosolutions/geoserver/rest/manager/GeoServerRESTFileSystemManagerTest.java
@@ -20,23 +20,31 @@ public class GeoServerRESTFileSystemManagerTest extends GeoserverRESTTest {
 
     @Test
     public void testXmlConfigExists() {
+        if(!enabled()) return;
+        
         assertTrue(manager.getFileSystemResourceManager().exists("global.xml"));
         assertFalse(manager.getFileSystemResourceManager().exists("global2.xml"));
     }
 
     @Test
     public void testXmlDownload() {
+        if(!enabled()) return;
+        
         assertTrue(manager.getFileSystemResourceManager().download("global.xml").contains("claudius.ptolomaeus@gmail.com"));
         assertNull(manager.getFileSystemResourceManager().download("global2.xml"));
     }
 
     @Test
     public void testXmlDownload2() {
+        if(!enabled()) return;
+        
         assertNotNull(manager.getFileSystemResourceManager().download("data/nyc/poi.shx"));
     }
 
     @Test
     public void testUploadAndDeleteFile() {
+        if(!enabled()) return;
+        
         manager.getFileSystemResourceManager().delete("testfile.txt");
         
         assertFalse(manager.getFileSystemResourceManager().exists("testfile.txt"));
